+++ commons/src/commonTest/kotlin/com/vitorpamplona/amethyst/commons/model/nip30CustomEmojis/OwnedEmojiPackTest.kt
@@ -18,7 +18,7 @@
  * AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package com.vitorpamplona.amethyst.model.nip30CustomEmojis
+package com.vitorpamplona.amethyst.commons.model.nip30CustomEmojis
 
 import com.vitorpamplona.quartz.nip30CustomEmoji.EmojiUrlTag
 import org.junit.Assert.assertEquals
